Fixes honeypot_check_field and its action hook
Changes the conditions in wdhp_honeypot_check_field to check for the correct
array; it was originally checking an index of an array which didn't exist.
Also adds a check to make sure the array index exists.

Changes the action hook from wp_insert_comment to comment_post. The former
was incorrectly used to retrieve the honeypot value. The WP_Comment class
only contains pre-determined properties, it does not output custom fields
as part of its properties. Therefore, $_POST array is now used instead.