Changes honeypot stratgey
The previous strategy used was ineffective because it added a honeypot
input with a unique id and name. If a spam bot targets the name, email,
url and comment fields by id or name, it will never fill in the honeypot
field. However, by duplicating the comment field, setting the original as
a trap, and using the duplicate with a unique name as the new legitimate
field, we can trick more spam bots.